Implemented Cong's changes.

+++ back-end/startup.py
@@ -19,33 +19,42 @@ def main():
     # parser.add_argument("-V", "--video", action="store_true")
 
     args = parser.parse_args()
+    if True not in vars(args).values():
+        parser.print_help()
+        return
 
     # The testing flag is only applied to the telemetry server. This is
     # appropriate because the test webpage only supports the telemetry server.
-    tel_args = ["python", "server.py", "-T"]
+    tel_args = ["python2", "server.py", "-T"]
     if args.test:
-        tel_args = ["python", "server.py", "-t", "-T"]
+        tel_args = ["python2", "server.py", "-t", "-T"]
     if args.aprs:
-        a_pid = Popen(["python", "server.py", "-A"])
+        a_pid = Popen(["python2", "server.py", "-A"])
     if args.telemetry:
         t_pid = Popen(tel_args)
     # if args.video:
-    #     v_pid = Popen(["python", "server.py", "-V"])
+    #     v_pid = Popen(["python2", "server.py", "-V"])
 
     try:
         # This loop checks if each subprocess is still running. If any of the
         # subprocesses have stopped, then they're restarted.
         while True:
-            if a_pid.poll() is not None:
-                a_pid = Popen(["python", "server.py", "-A"])
-            if t_pid.poll() is not None:
-                t_pid = Popen(tel_args)
-            # if v_pid.poll() is not None:
-            #     v_pid = Popen(["python", "server.py", "-V"])
+            if args.aprs:
+                if a_pid.poll() is not None:
+                    a_pid = Popen(["python2", "server.py", "-A"])
+            if args.telemetry:
+                if t_pid.poll() is not None:
+                    t_pid = Popen(tel_args)
+            # if args.video:
+            #     if v_pid.poll() is not None:
+            #         v_pid = Popen(["python2", "server.py", "-V"])
     except KeyboardInterrupt:
-        a_pid.terminate()
-        t_pid.terminate()
-        # v_pid.terminate()
+        if args.aprs:
+            a_pid.terminate()
+        if args.telemetry:
+            t_pid.terminate()
+        # if args.video:
+        #     v_pid.terminate()
         return
 
 if __name__ == "__main__":
